make docs build successfully , activate dochack (#426)
* make docs more compliant with rst

* fix documentation in header of einsum

* fix openmp import path in `blas_l3_gemm_macro_kernel.nim`

* fix type import in nnpack_interface.nim

* make all `noInit` pragmas consistent

Well, nim still complains. All `noInit` pragmas should be the other
way round I fear. :|

* add missing import for `auxiliary_lapack` "main module"

* [laser] disallow `gemm_packed.nim` main proc, not compatible

I assume the `when isMainModule` parts are straight from
laser. However, the tensor definitions are missing / in different
files in arraymancer and thus the code is broken.

* add code to generate nimdoc.cfg and documentation

* add updated nimdoc.cfg

* remove stray `p` from html template

We don't want every page to have a little 'p' in the bottom left
corner, do we?

* compare and emit names w/o `.html`

* add `Other docs` category and fix dir names of ag, la

* [gemm legacy] remove not used import, fix used import path

* merge nimdoc creationg into `docs', add `templates` file

* correctly set path to `source` and `edit` in docs

* add scrolling for the header

Currently setting the max size to 800 px. I'm not a css wizard, so I
don't know how to make it take into account device size.

* fix naming for layer initializations

* update nimdoc.cfg

* move doc generation to src/docs

If the files are located outside the src dir, nimble will fail to
install the package, since for some reason it does someting with its
generated nim script file from the ~/.nimble/pkg directory from which
it obviously cannot find docs / docs, since that isn't part of the
source.

* move docs back to docs/, hack around actual problem

Sorry to solve it this way. I feel like this might be a nimble
bug. Even if the files are in the actual source directory of the
package, it still says it can't import them.