add kernel density estimation (KDE) (#444)
* add draft for kernel density estimation (KDE)

* move `linspace`, `logspace` to `init_cpu`, add tests

* move sort, sorted to new `algorithms.nim`, add tests

* move percentile, iqr to aggregate, add tests

* remove map_inline dep in init_cpu to avoid recursive import

* move `gauss` to new `stats/distributions.nim`, add tests

* move other kernel distributions to `distributions.nim`, add tests

The actual kernels, which take x, x_i, bw are now generated with a
template to provide the distributions we use also as general
functions. They may be useful outside of the scope of KDE after all.

* clean up `kde.nim` and fix a few bugs in the calculation

* fix paths for stats tests

* fix fancy indexing for access (in)directly via tensor, seq

* add `argsort`, required for weight support for KDEs

When implementing support for weights for the KDE we need to be able
to sort two tensors according to the sorting of one tensor. `argsort`
can be used for just that.

* allow to sample at custom points by handing tensor, seq as `samples`

* add a few simple tests for KDE, comparison with sklearn

* add tests for epanechnikov, box

* implement weights support for kde

* only export nimIdentNormalize on devel

* fix typo, convert to float

* make act. `kde` take `static KernelFunc` to allow inline to work

Use a template in convenience `kde` to call `kde` with a static argument.

Co-authored-by: Mamy Ratsimbazafy <mamy_github@numforge.co>