move allocator logic to type definition, define CpuStorageObj
The finalizer has to be defined in the same file as the type
definition for ARC (even though we don't use the finalizer for ARC,
but the actual destructor). Not sure if this applies to the
destructor too, but I assume?

Also have to define an actual object from which a `ref object` is
created for the signature of the `=destroy` proc.