replace `supportsCopyMem` by custom typeclass useable in type section
Instead of using two different methods within the type section and in
the code, which can lead to certain objects appearing in different
branches in the type section and code (e.g. variant objects are not
mem copyable, but are not in `not (T is string and T is ref)`, we use
a custom type class for both.