make `setZero` underlying seq at Tensor.size, do not call in newTensor
`reset` sets the length of the underlying sequence to 0! This way we
actually reset it to empty and resize to tensor size, which should
reallocate a new block that will be zeroed for us.
Since `newSeq` already zeroes in the first place, we don't have to
call `setZero` when constructing a seq based tensor.