Laser tensor PR rebase (#477)
* Lay out the base data structure and iteration from laser

* Replace the core tensor data structure by Laser and see what breaks

* Workaround + progress. Now compiler quits in the middle of "nim c -o:build/tests_tensor_part01 -r tests/_split_tests/tests_tensor_part01.nim"

* remove usage of dataArray

* rebase merge leftover

* Handle Nim distinctBase move to typetraits in devel

* supportsCopyMem in 1.0.6

* Small fixes on deprecated "data" proc

* Make the test suite compile (returning var seq is broken)

* Fix offset bug when iterating

* deprecated cleanup

* fix out-of-bounds test

* typetraits is always needed

* deprecate toRawSeq (cannot be made backward compatible)

* The default LASER_MAX_RANK is 6

* Fix sparse softmax cross-entropy test. Nim random is inclusive :/

* Fix the new parallel copyFrom (impacts index_select/embedding/GRU). Passing all tests with gc:markandsweep

* Update changelog + unsafe_raw_data leftovers

* Make laser-tensor-iterator compile with latest Nim

* update selectors to work with laser

* Trying to narrow down the tests that trigger a refcounting bug at a

later state:
- test_display
- test_higherorder
- test_ufunc
Commonalities:
- all rely on higher order templates map/apply
- all have at least a test that involve mapping a string
Issue can be reliably detected in its own module by wrapping tests into a proc and avoid globals and then calling GC_fullcollect or GC_collectZct

* Fix the bug left (https://github.com/nim-lang/Nim/issues/13598) and reactivate full test suite

* Don't use deprecated .data accessor + the mutable one is broken so deleted without deprecation

* use unsafe_raw_offset in example instead of get_offset_ptr

* move laser files to tensor subdirectory

This is needed, due to the changes that happened to the Arraymancer
directory layout in the meantime.

* add MetadataArray, fix imports for typetraits, ast_utils

* add lost `replaceNodes`, export Tensor

* fix broken `init_cpu` rebase, add linspace, logspace

* fix `isMemOwner` field name

* export full `datatypes`, remove duplicate rank, size

The latter procs are defined in the laser part.

* also export dynamic_stack_arrays

* remove `p_init_cpu`

* fix p_shapeshifting imports

* remove old metadataArray import

* fix import paths for tensor submodule

* remove old code from bad merge

* remove duplicate `is_C_contiguous`

* fix bad merge of data_structure accesses

* fix more tensor imports

* fix more imports

* fix sorting impl for laser tensors

* fix hopefully final import

* make some funcs procs due to impl changes

* move allocator logic to type definition, define CpuStorageObj

The finalizer has to be defined in the same file as the type
definition for ARC (even though we don't use the finalizer for ARC,
but the actual destructor). Not sure if this applies to the
destructor too, but I assume?

Also have to define an actual object from which a `ref object` is
created for the signature of the `=destroy` proc.

* remove a self assignment in pca.nim

This self assignment breaks the code for usage with arc/orc.

For some reason this seems to call the destructor of the Tensor?

* remove imports of now removed allocator.nim

* call `=destroy` for seq based fallback tensor buffers

* replace `supportsCopyMem` by custom typeclass useable in type section

Instead of using two different methods within the type section and in
the code, which can lead to certain objects appearing in different
branches in the type section and code (e.g. variant objects are not
mem copyable, but are not in `not (T is string and T is ref)`, we use
a custom type class for both.

* add complex types to mem copyable type class

* perform check on shape instead of size for `copyFrom`

The docstring mentions the shapes of the input tensors should
match. For contiguous mem copyable tensors only the size is
checked (because it is technically enough in the case)

* fix `test_filling_data` to actually use same shape

* remove {.noInit.} from `asType` to work around Nim #16253

* special case `atIndexMut` b/w ARC / KnownSupportsCopyMem / seq tensors

* add a `newTensorUninit` overload for 1D tensor to workaround bug

As mentioned in the docstring, this is an overload, because calling
the varargs version is broken if being called from a template.

* remove `data` proc again after readded by mistake

This was already removed due to possibly being bugged in
https://github.com/mratsim/Arraymancer/pull/420/commits/56de8c9d56555173ce5c175997957f66e3717588

* fix when condition in `setZero`

* make `setZero` underlying seq at Tensor.size, do not call in newTensor

`reset` sets the length of the underlying sequence to 0! This way we
actually reset it to empty and resize to tensor size, which should
reallocate a new block that will be zeroed for us.
Since `newSeq` already zeroes in the first place, we don't have to
call `setZero` when constructing a seq based tensor.

* set `memalloc` pointer to nil after freeing in finalizer / destroy

* define `=` as {.error.} for `CpuStorageObj` for ARC/ORC

* change `toRawSeq` to not use removed `data`, copy manually

This at least _seems_ to work fine as far as I can test, both using
ARC and normal GC.

`export_tensor` still used `data` as well. For the time being we can
use `toRawSeq` until a better solution is in place.

* add a test case checking for self assignment bug

As long as this test case passes, self assignment of objects
containing Tensors using ARC is broken in arraymancer!

* add nimble tasks to test with ARC, use for travis / appveyor

* make ARC specific test only run under ARC

* remove ARC test on appveyor since test runs Nim 1.0.6

* add bool to KnownSupportsCopyMem

* make `unsafe_raw_*` procs only valid for KnownSupportsCopyMem

* make `atIndex(Mut)` only use unsafe_raw procs if not KnownSupportsCopyMem

* make `*stridedIteratior` templates work for not KnownSupportsCopyMem

* make `atContiguousIndex` not use `unsafe_raw` procs for non mem copyable

* avoid `forEach` macro for non KnownSupportsCopyMem in newTensorWith

* make `*_inline` templates work for non KnownSupportsMemCopy

* add more safety checks for map/apply reg KnownSupportsCopyMem

* error for `get_*_ptr`, `dataArray` for not KnownSupportsCopyMem

* only allow `argmax_max` for `SomeNumber` since it uses unsafe_raw

* do not auto import `nimhdf5` if available anymore

- `io_hdf5` is not imported automatically anymore if the module is
  installed. The reason for this is that the HDF5 library runs code in
  global scope to initialize the HDF5 library. This means dead code
  elimination does not work and a binary will always depend on the
  HDF5 shared library if the `nimhdf5` is installed, even if not used.

* add `fromBuffer` to create a tensor w/o copy from a raw pointer

* Remove MetadataArray ambiguous call, pass OpenCL tests, can compile Cuda for Nim 1.0.x

Co-authored-by: Mamy André-Ratsimbazafy <mamy_github@numforge.co>