Fix doc generation due to change in src layout, fix `csize_t` regression (#482)
* delete deprecated file (superseeded by laser/private/nested_*)

* fix code using csize_t by converting manually

`csize_t` is an unsigned int. The height / width variables here are
normal ints though. Given that there won't be negative heights /
widths converting to csize_t shouldn't cause any issues.

* fix import of `ast_utils` in foreach_staged

* fix comments in `nnpack`, cause `new section expected` err otherwise

* fix headlines in rst files, caused `new section expected`

* add note to `docs.nim`, cause I even managed to confuse myself

* fix path used for gen_docs to generate correct path

* update autogenerated nimdoc.cfg

Note: this file doesn't really have to be checked in at all. But given
that it is, I'd rather update it to reflect the state of the repo.