play nim compile type-whack-a-mole
We don't want to export `<` for tensors that does a element wise
comparison until something is smaller or larger in either argument.

As I couldn't make the compiler happy with a `bind `<`` in the code,
let's work around it by creating a `distinct` type that we only use in
the heap queue! :partying_face: