clean up and modify distances to accomodate kdtree
We need to be able to compute the pairwise distance between a single
point and a tensor of points. Hence the new logic in
`pairwiseDistances`.

Note: we could invert the whole logic of `distance`
vs. `pairwiseDistance` so that `pairwiseDistance` is the "main"
procedure that computes distances and then only sums along the
"feature/dimensionality" axis and keeps everything along the
"observations" axis.

The code becomes pretty messy in that case though, as all the
`distance` procs have to do unsqueezing, squeezing and extracting
elements from rank 1 tensors with 1 element.

In the future one may want to revisit this from a performance
standpoint, but the code is clearer as it is imo.