better least sq. docs, make `rcond` argument, fix workspace calc
The workspace calculation was flawed. For certain large arguments the
used 25 as the smallest workspace leaf was not enough, resulting in
errors on calls.

Also make `rcond` changeable by user. It's a tricky input, because it
may result in wrong results if not set correctly. The previous value
yielded wrong results for certain problems (Savitzky-Golay filter in
scinim / ggplotnim).