DBSCAN rebase using k-d tree (#518)
* Add dbscan clustering method

* Bugfix: compute jaccard distance instead of similarity

* move files to new sub directory structure

* clean up ml imports

* fix imports according to current dir structure

* add `toCopy` argument to `argsort` to avoid modifying argument

Not sure if there should instead just be a `argsorted` proc. Hmm.

* use tensor argsort instead of custom seq argsort

* [aggregate] `cumsum`, `cumprod` now assume default axis 0

Allows calling them without explicitly stating an axis.

* [distances] change metric code to typedesc[T] calling convention

Also add a bunch of doc strings. :)

* rename distance.nim to distances.nim

* fix imports / exports according to distances.nim

* [algorithms] fix type of mutable local variable

* fix imports in `dbscan`

* [kmeans] use `distances.nim` and `cumsum` from aggregate

Less duplication, yay

* [kmeans] fix wrong assert in kmeans, may be more clusters than dims

These assertions don't make a lot of sense to me. `n_cols` is the
dimensionality of the data. `n_clusters` is the desired number of
clusters kmeans should find. There may be many more clusters than
dimensions of course. Perfectly fine to have 1000 clusters in a 2D space.

* [neighbors] use k-d tree by default for nearest neighbor computation

* [dbscan] allow custom metrics in dbscan & add some docstring

* [neighbors] add TODO

* clean up and modify distances to accomodate kdtree

We need to be able to compute the pairwise distance between a single
point and a tensor of points. Hence the new logic in
`pairwiseDistances`.

Note: we could invert the whole logic of `distance`
vs. `pairwiseDistance` so that `pairwiseDistance` is the "main"
procedure that computes distances and then only sums along the
"feature/dimensionality" axis and keeps everything along the
"observations" axis.

The code becomes pretty messy in that case though, as all the
`distance` procs have to do unsqueezing, squeezing and extracting
elements from rank 1 tensors with 1 element.

In the future one may want to revisit this from a performance
standpoint, but the code is clearer as it is imo.

* [kdtree] allow usage of any metric in k-d tree query procedures

* [neighbors] use custom metric in `nearestNeighbor`

* [dbscan] add docstrings for DBSCAN, simplify one line

* fix `distances` bugs and add test cases

* export sets before nim version 1.4 as `bind` doesn't work there

* [tests] add test case for DBSCAN, another for kmeans

And mini improvement for `neighbors`

* update changelog

* fix name of DBSCAN test

* make sure to run clustering test with old RNG

Co-authored-by: abieler <andre.bieler81@gmail.com>