fix underlying issue of #523 by assigning to correct index
`enumerate` yields the contiguous index (it *enumerates* all elements
from 0 to t.size). Thus for rank N > 1 tensors accessing with `[i]`
causes a runtime error, as the index is not N elements.

For the memcopyable types, this is not an issue here, as it uses
`unsafe_raw_offset` to access the pointer directly.