Fix issue #523 by fixing `map_inline` for non memcopyable rank > 1 (#524)
* fix underlying issue of #523 by assigning to correct index

`enumerate` yields the contiguous index (it *enumerates* all elements
from 0 to t.size). Thus for rank N > 1 tensors accessing with `[i]`
causes a runtime error, as the index is not N elements.

For the memcopyable types, this is not an issue here, as it uses
`unsafe_raw_offset` to access the pointer directly.

* [tests] add tests for issue #523

* [tests] add test case for `map_inline` using slicing

* [tests] take out MNIST test as it is too flaky for a CI

It's a good idea to make sure the download works, but it's not really
suitable for a CI if it fails so damn often.