Fix more undeclared identifier issues (#529)
* restrict `rank` in `data_structure` to CUDA/CL, add laser `rank/size`

Instead of trying to have one definition for these, it's easier to
just have a separate impl, as these procs are so simple anyway. This
let's us avoid having to do more imports that break things... :/

* remove unnecessary import in `higher_order_applymap`

This import can lead to a recursive import that causes identifier
resolution issues.

* explicitly call `pairs` iterator

For reasons unknown to me, if we do not specify the `pairs` iterator
here, the iterator is not found. We can neither `bind` in local nor
`mixin` for calling. This works though...

* without importing higher_order_applymap, `apply2_inline` won't be visible

Unfortunately again, we cannot `mixin` the identifier for later
resolution. It just doesn't work. Fortunately, the recursive import
this brings to the table, could be resolved, because the
higher_order_applymap file doesn't have to import the accessors write file.

* fix possible undeclared ident `|` issue by explicitly creating object

Instead of hoping the resolution of the rather trivial
```
slice .. slice | 1
```
works (mixin again doesn't work), we just construct the object this
will be resolved into directly.

* add semi related test for slices in generics

This test unfortunately doesn't directly represent the numericalnim
issue this PR should fix. Still better to have it than not

* update changelog

* fix test case for slicing in generics