extend `einsum` to work in generic contexts
In a generic the `[]` is treated as a `nnkCall` that is of an open sym
choice. This breaks the current einsum logic. This commit makes it
work. However, specifically `nnkClosedSymChoice` is still not
supported. I don't think it should appear here, or rather I'd like a
test case before I add it.