do not compile with AVX512 support by default
This commit fixes issue #505, which is a codegen issue arising from
generated C code that is only valid if the `-mavx512dq` compilation
flag is handed to the C compiler.

The issue is that:
- we cannot generate the code without the compilation flag (and
compile it successfully)
- handing the compilation flag makes the code compile, but forces
every CPU to attempt to run the AVX512 code, which is an illegal
operation for CPUs that do not support it.

For this reason AVX512 support is hidden behind a `-d:avx512`
compilation flag that needs to be handed. The resulting binary then
*will* use AVX512 for gemm.