fix `einsum` in generics more elegantly by rewriting AST into untyped
Given that we have a very small amount of possible AST within the
context of the `einsum` DSL, rewriting the AST from a typed open sym
choice form back to an untyped form is pretty sane.

Includes a longer example of what the AST transformation looks like.

Removes the logic that was added in #539 that made the macro work with
typed AST.

Also includes an error message if the `findIdx` helper procedure fails
to find a match (which should never happen, but can during development
work on the macro logic).