Update docs, add GH action for docs, fix a few styleCheck hints (#556)
* fix some styleCheck hints

In particular the `reduce_inline` call in `min` is weird. It seems the
compiler confuses the argument and generic type and thus throws a hint.

* [docs] fix up doc generation for current `nim doc` behavior

* [docs] fix up doc strings/RST files where nim doc now complains

* move `nn_dsl` into `nn` subdir to conform to docgen logic

nn_dsl was now the only file in the source root directory that
actually contained actual source

* [docs] update docgen code for `nn_dsl`

* update GitHub action to also build docs

* update local `nimdoc.cfg`

Note: we should be generating this file from now on via GH actions. If
that works, we can remove this file from the repo.

* build docs on version 1.6 branch as we don't test devel