Improve the display of rank 2 column Tensors
Fix an issue when representing the first element of rank 2 single column tensors (the first element was not being indented properly, and the left side "line" was not being "drawn").

There is still an issue when displaying the 2D tensors that have more than one column. The first column is not properly aligned. This is not fixed by this commit.