k-d tree performance improvements & KDE `adjust` fix (#587)
* significantly improve kd-tree performance (special case euclidean)

This adds a special code path for the euclidean distance
calculation (and removes the need to squeeze input data).

As the default euclidean distance is very common it's worthwhile to
have a branch that is >2x faster

* [kde] actually apply `adjust` if given to adjust auto bandwidth

* [kdtree] avoid more expensive `pow` calls if squared (p = 2)

* avoid axis iterator in nearest neighbor

* [experimental] replace HeapQueue by a sorted seq

The sorted seq is implemented such that the `pop` operation is
trivial (last element + setLen(N-1)). In my use case it seems to give
another big performance boost.

* update changelog

* use atomic counter for `nonzero` for Nim > 1.6

The code uses a local variable outside the scope of `map_inline`. If
compiled with OpenMP support that local variable won't be modified as
each thread's code gets their own counter, breaking the code.

* fix import of `random` in `init_cpu`

* fix import of math, random, table sin `kmeans`

* fix import of random in openmp

* to avoid sorted seq dep, use HeapQueue

In my application the sorted seq is a bit faster, but for now let's
use the stdlib heapqueue further.

Will benchmark the difference again and maybe introduce submodule if
performance difference large enough.

* [distances] use raw buffers directly in squared Euclidean if contig.

If the input tensors are contiguous use the direct buffers
directly. This way we can ignore the potential rank 1 vs rank 2
issues.

Otherwise we fall back to the old implementation.

* [distances] handle unsqueezing a potential [n_dimensions] input tensor

So that one input can be handed as a pure rank 1 tensor.