Add support for combining span slices with negative steps.
This adds support for `a[_|-2]` and `a[_.._|-2]` syntax, both of which are syntactic sugar for `a[^1..0|-2]`. That is, we assume that the intention behind a "full" span combined with a negative step is to go through the tensor in reverse order.

This makes is extremelly simple to reverse a tensor (i.e. `a[_|-1]`) and is equivalent to python's `a[::-1]` syntax.