Add support for combining span slices with negative steps. (#598)
* Add support for combining span slices with negative steps.

This adds support for `a[_|-2]` and `a[_.._|-2]` syntax, both of which are syntactic sugar for `a[^1..0|-2]`. That is, we assume that the intention behind a "full" span combined with a negative step is to go through the tensor in reverse order.

This makes is extremelly simple to reverse a tensor (i.e. `a[_|-1]`) and is equivalent to python's `a[::-1]` syntax.

* Update and expand the slicing tutorial

Adds an explanation of the use of the new negative spans to reverse a tensor as well as some additional examples.

* Add an example of reversing a tensor using a negative span to the README

Also fix a couple of markdown warnings.

* Fix a couple of typos in comments

* Fix a few markdown warnings in the README

* Ensure that all the examples in the README compile and that their outputs match the README.

* Improve the documentation of the -d:blas and -d:lapack settings and the installation of BLAS and LAPACK on windows.

* Fix incorrect tensor initialization in README example.

* Fix small typo in comment

---------

Co-authored-by: Angel Ezquerra <angel_ezquerra@keysight.com>