Add some padding to the "Technical reference" submenu to make it easier to use
The "Technical reference" submenu is horizontal which with the default padding makes it quite hard to use.
It is very easy to move the mouse outside of the menu and then the menu is closed.
Adding some padding makes it easier to use. The drawback is that it looks a little different than the other menus, which are vertical.

The ideal solution would be to make all the menus vertical, but since the "Technical reference" submenu is 2 levels deep this is not straightforward to do.