Add at_mut template to enable masked_fill on slices
This can be useful, for example, when assigning a value into a chain of slice operations which are usually considered immutable even if the original tensor is mutable. For example, the following code:

```nim
var x = arange(20).reshape([4, 3])
x[1..2, _][condition] = 100
```
will fail with a `a slice of an immutable tensor cannot be assigned to` error.
Using `at_mut` you can do:
```nim
x.at_mut(1..2, _)[condition] = 100
```
which works fine.