Add moveaxis procedure
This is also missing from numpy. It is similar to `permute` but instead of having to specify the whole reordered list of axis you can simply set an initial axis and its target position which is sometimes more convenient.

This commit also adds tests for the `moveaxis` and `permute` procedures. It also improves the documentation of the `permute` procedure.