Add flatten function
`flatten` flattens a tensor, returning a rank-1 tensor with the same data as the input.

This is the same as `t.reshape([t.size.int])`. Therefore, if possible no data copy is done and the returned tensor shares data with the input.
If input is not contiguous, this is not possible and a copy will be made.