Add roll support for tensors of rank > 1
This adds two roll types (which are the same for rank 1 tensors, but different for higher rank tensors): "global" (axis-less) roll and "axis" roll. The former flattens the input tensor before rolling the elements and then reshapes the rolled tensor back to the original shape, while the latter rolls the positions of _slices_ of a tensor taken along a specific axis.