Missing numpy functions (#619)
* Add documentation for the triu and tril procedures.

* Add tri() procedures and export existing triu() and tril() procedures.

Also tweaks the assert messages of the diagonal() proc a little to match similar messages in the new functions.

* Add tests for the tri(), tril() and triu() procedures.

Tests have been added to test_diag.nim which will be renamed to test_special_matrices.nim.

* Rename test_diag.nim to test_special_matrices.nim

* Add test_special_matrices to the test suite.

* Add geomspace procedure (which was missing from numpy).

Also adds messages to a couple of asserts on the linspace procedure.

* Add sinc function (which was missing from numpy).

Also adds a corresponding test.

* Add meshgrid function.

* Add all and any functions (which were missing from numpy).

* Replace references to the `future` module with references to the `sugar` module in all docstrings.

* Add flatten function

`flatten` flattens a tensor, returning a rank-1 tensor with the same data as the input.

This is the same as `t.reshape([t.size.int])`. Therefore, if possible no data copy is done and the returned tensor shares data with the input.
If input is not contiguous, this is not possible and a copy will be made.

* Fix typo on accessors/enumerateAxis docstring

* Add roll function for rank 1 tensors

Support for higher rank tensors will be added later.

* Add roll support for tensors of rank > 1

This adds two roll types (which are the same for rank 1 tensors, but different for higher rank tensors): "global" (axis-less) roll and "axis" roll. The former flattens the input tensor before rolling the elements and then reshapes the rolled tensor back to the original shape, while the latter rolls the positions of _slices_ of a tensor taken along a specific axis.