More missing numpy (#622)
* Add Scalar-Tensor element-wise comparison operators

We already supported Tensor-Scalar comparisons. This adds support for the reversed order (i.e. with the scalar first).

* Add append procedures

This adds 2 append procedures, one which appends the items of a tensor and one that appends the items of an open array. They are otherwise the same.

These new functions are similar to (but more limited than) numpy's append. The main differences are that:
- They require their tensor inputs to be of rank-1 (i.e. they require their inputs to be explicitly flattened if their rank is higher than 1).
- They do not support the axis argument.

IMHO the automatic flattening of the output of numpy's append can lead to very surprising behavior (hence limitation #1). Only supporting rank-1 tensors makes the axis argument unnecessary (hence limitation #2).

Users that need to work with higher ranked tensors can always use the existing concat procedure.

* Add moveaxis procedure

This is also missing from numpy. It is similar to `permute` but instead of having to specify the whole reordered list of axis you can simply set an initial axis and its target position which is sometimes more convenient.

This commit also adds tests for the `moveaxis` and `permute` procedures. It also improves the documentation of the `permute` procedure.

* Add a clone of numpy.vander

This also makes `vandermonde` more flexible (make it possible to set a tensor of orders, or to not provide an order or an input tensor).

Finally, this simplifies the slicing tutorial a little, by using the new generalized vandermonde procedure instead of manually creating the input vandermonde matrix.

* fix minor typo

---------

Co-authored-by: Vindaar <basti90@gmail.com>