Add a single tensor version of the complex() procedure.
Note that you while can also convert a real tensor into a complex tensor by means of the `asType` procedure, this has the advantage that you can use the same function to combine a real and imaginary tensor or a single real tensor into a complex tensor.

Another advantage is that this function will automatically use the right Complex type for the output tensor, leading to more generic code.

That being said, `asType` can still be useful when you want to control the type of the Complex tensor.