Some warning removals (#635)
* Remove "A custom '=destroy' hook which takes a 'var T' parameter is deprecated" warning from tensor\datatypes.nim

Note that there is still another similar warning left in:

\laser\primitives\matrix_multiplication\gemm_tiling.nim(321, 3)

* Fix typo in comment

* Fix several `imported and not used [UnusedImport]` warnings

In particular:
- Warning: imported and not used: 'datatypes'
- Warning: imported and not used: 'complex'
Warning: imported and not used: 'math'

* Remove a ` conversion from string to itself` hint

Removed from `examples\ex06_shakespeare_generator.nim`.

* Remove `Deprecated since 1.5; TaintedString is deprecated` warning

* Do not crash when no argument is passed to ex06_shakespeare_generator

* Disable the remaining `'=destroy' hook which takes a 'var T' parameter is deprecated` warning

We cannot really fix that warning because it is due to the fact that (as for 2.0.2) nim creates a var destructor when `new X, T` is called. Since we cannot fix it we simply disable it.

* Normalize all std imports to use `import std / xxx`

This also removes a few unnecessary imports of the complex and math modules (which are automatically exported by arraymancer). The test touches a lot funtions but the changes are minimal and simple (basically changing things like `import unittest, os` into `import std / [unittest, os]` and so on).

* Remove `imported and not used: 'sequninit' [UnusedImport]` warning

This warning only happens with devel. The fix is to make sequninit.nim "used" when newSeqUninit is not needed.