properly gensym shim templates in p_accessors (#642)
For code like:

```nim
when foo():
  let data = ...
else:
  template data: untyped = ...
```

in a template, since the `let data` is automatically gensym'd, the `template data` must also be marked as `gensym`, since routines are injected by default in templates. Due to a bug in the Nim compiler (https://github.com/nim-lang/Nim/issues/23326) the `template data` here was forced to being gensym due to the previous `let data` symbol. This behavior might not change, but the `template data` can be marked as gensym anyway for clarity.