Missing special matrices (#647)
* Add `circulant` special matrix

A circulant matrix is a square matrix in which each column (or row) is a cyclic shift of the previous column (or row). This is missing from numpy / scipy.

* Add way to construct hankel matrices

A Hankel matrix has constant anti-diagonals, with c as its first column and r as its last row (note that `r[0]` is ignored but _should_ be the same as `c[^1])`. This was missing from numpy / scipy.

* Add way to construct toeplitz matrices

A Toeplitz matrix has constant diagonals, with c as its first column and r as its first row (note that r[0] is ignored but should be the same as c[^1]). This was missing from numpy / scipy.

There is also a single input version.