Add `unique` functions to algorithms.nim
These functions are similar to but not as fully featured as `numpy.unique`.
They are missing a way to count the number or returning the indexes of the unique elements.
However, they make it possibel to (optionally) sort the output, or to use a more efficient algorithm if the input is already sorted.